Redirect to app, successfuly authenticate
